dh_backtest/requirements.txt
arrow==1.3.0
dash==2.17.1
dash_bootstrap_components==1.6.0
ib_insync==0.9.86
numpy==2.0.1
pandas==2.2.2
plotly==5.23.0
python_dateutil==2.9.0.post0
pytz==2024.1
termcolor==2.4.0
